feat(libp2p): platform New() shim (native transports / wasm wasmws)
Keep the Service struct and all methods shared in libp2p.go; move the
native constructor (TCP/WS/WSS, AutoTLS, AutoNAT) to libp2p_native.go and
add libp2p_js.go which builds a browser host over the wasmws WebSocket
transport with no listeners. Adds js errno sentinels and the wasmws dep.
