+++ pkg/p2p/libp2p/libp2p_js.go
@@ -1,0 +1,234 @@
+//go:build js
+
+// Copyright 2020 The Swarm Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+
+package libp2p
+
+import (
+	"context"
+	"fmt"
+	"net"
+
+	ocprom "contrib.go.opencensus.io/exporter/prometheus"
+	"github.com/ethersphere/bee/v2/pkg/addressbook"
+	"github.com/ethersphere/bee/v2/pkg/bzz"
+	beecrypto "github.com/ethersphere/bee/v2/pkg/crypto"
+	"github.com/ethersphere/bee/v2/pkg/log"
+	m2 "github.com/ethersphere/bee/v2/pkg/metrics"
+	"github.com/ethersphere/bee/v2/pkg/p2p"
+	"github.com/ethersphere/bee/v2/pkg/p2p/libp2p/internal/blocklist"
+	"github.com/ethersphere/bee/v2/pkg/p2p/libp2p/internal/breaker"
+	"github.com/ethersphere/bee/v2/pkg/p2p/libp2p/internal/handshake"
+	"github.com/ethersphere/bee/v2/pkg/storage"
+	"github.com/ethersphere/bee/v2/pkg/swarm"
+	"github.com/ethersphere/bee/v2/pkg/topology/lightnode"
+	"github.com/ethersphere/bee/v2/pkg/tracing"
+	"github.com/libp2p/go-libp2p"
+	"github.com/libp2p/go-libp2p/core/crypto"
+	"github.com/libp2p/go-libp2p/core/network"
+	"github.com/libp2p/go-libp2p/core/protocol"
+	basichost "github.com/libp2p/go-libp2p/p2p/host/basic"
+	"github.com/libp2p/go-libp2p/p2p/host/peerstore/pstoremem"
+	rcmgr "github.com/libp2p/go-libp2p/p2p/host/resource-manager"
+
+	wasmws "github.com/v1rtl/go-libp2p-wasmws"
+)
+
+// New constructs a libp2p Service for wasm/browser builds. Unlike the native
+// build, there are no TCP/WS listeners, no AutoTLS, and no AutoNAT: the browser
+// can only dial out, and it does so over the wasmws (browser WebSocket)
+// transport.
+func New(ctx context.Context, signer beecrypto.Signer, networkID uint64, overlay swarm.Address, addr string, ab addressbook.GetPutter, storer storage.StateStorer, lightNodes *lightnode.Container, logger log.Logger, tracer *tracing.Tracer, o Options) (s *Service, returnErr error) {
+	logger = logger.WithName(loggerName).Register()
+
+	security := libp2p.DefaultSecurity
+	libp2pPeerstore, err := pstoremem.NewPeerstore()
+	if err != nil {
+		return nil, err
+	}
+
+	if o.Registry != nil {
+		rcmgr.MustRegisterWith(o.Registry)
+	}
+
+	_, err = ocprom.NewExporter(ocprom.Options{
+		Namespace: m2.Namespace,
+		Registry:  o.Registry,
+	})
+	if err != nil {
+		return nil, err
+	}
+
+	// Tweak certain settings
+	cfg := rcmgr.PartialLimitConfig{
+		System: rcmgr.ResourceLimits{
+			Streams:         IncomingStreamCountLimit + OutgoingStreamCountLimit,
+			StreamsOutbound: OutgoingStreamCountLimit,
+			StreamsInbound:  IncomingStreamCountLimit,
+		},
+	}
+
+	limits := cfg.Build(rcmgr.InfiniteLimits)
+	limiter := rcmgr.NewFixedLimiter(limits)
+
+	str, err := rcmgr.NewStatsTraceReporter()
+	if err != nil {
+		return nil, err
+	}
+
+	rm, err := rcmgr.NewResourceManager(limiter, rcmgr.WithTraceReporter(str))
+	if err != nil {
+		return nil, err
+	}
+
+	var natManager basichost.NATManager
+
+	opts := []libp2p.Option{
+		libp2p.NoListenAddrs,
+		security,
+		// Use dedicated peerstore instead the global DefaultPeerstore
+		libp2p.Peerstore(libp2pPeerstore),
+		libp2p.UserAgent(userAgent()),
+		libp2p.ResourceManager(rm),
+	}
+
+	if o.NATAddr == "" && o.NATWSSAddr == "" {
+		opts = append(opts,
+			libp2p.NATManager(func(n network.Network) basichost.NATManager {
+				natManager = basichost.NewNATManager(n)
+				return natManager
+			}),
+		)
+	}
+
+	if o.PrivateKey != nil {
+		myKey, _, err := crypto.ECDSAKeyPairFromKey(o.PrivateKey)
+		if err != nil {
+			return nil, err
+		}
+		opts = append(opts,
+			libp2p.Identity(myKey),
+		)
+	}
+
+	transports := []libp2p.Option{}
+
+	var tcpResolver handshake.AdvertisableAddressResolver
+	if o.NATAddr != "" {
+		r, err := newStaticAddressResolver(o.NATAddr, net.LookupIP)
+		if err != nil {
+			return nil, fmt.Errorf("static nat: %w", err)
+		}
+		tcpResolver = r
+	}
+
+	var wssResolver handshake.AdvertisableAddressResolver
+	if o.EnableWSS && o.NATWSSAddr != "" {
+		r, err := newStaticAddressResolver(o.NATWSSAddr, net.LookupIP)
+		if err != nil {
+			return nil, fmt.Errorf("static wss nat: %w", err)
+		}
+		wssResolver = r
+	}
+
+	// In the browser the only available transport is the WebSocket transport
+	// backed by the host's WebSocket API.
+	transports = append(transports, libp2p.Transport(wasmws.New))
+
+	opts = append(opts, transports...)
+
+	if o.hostFactory == nil {
+		// Use the default libp2p host creation
+		o.hostFactory = libp2p.New
+	}
+
+	h, err := o.hostFactory(opts...)
+	if err != nil {
+		return nil, err
+	}
+
+	// Support same non default security and transport options as
+	// original host.
+	dialer, err := o.hostFactory(append(transports, security)...)
+	if err != nil {
+		return nil, err
+	}
+
+	if o.HeadersRWTimeout == 0 {
+		o.HeadersRWTimeout = defaultHeadersRWTimeout
+	}
+
+	handshakeService, err := handshake.New(signer, newCompositeAddressResolver(tcpResolver, wssResolver), overlay, networkID, o.FullNode, o.Nonce, newHostAddresser(h), o.WelcomeMessage, ab, h.ID(), o.ChequebookVerifier, logger)
+	if err != nil {
+		return nil, fmt.Errorf("handshake service: %w", err)
+	}
+
+	// Create a new dialer for libp2p ping protocol. This ensures that the protocol
+	// uses a different set of keys to do ping. It prevents inconsistencies in peerstore as
+	// the addresses used are not dialable and hence should be cleaned up. We should create
+	// this host with the same transports and security options to be able to dial to other
+	// peers.
+	pingDialer, err := o.hostFactory(append(transports, security, libp2p.NoListenAddrs)...)
+	if err != nil {
+		return nil, err
+	}
+
+	peerRegistry := newPeerRegistry()
+	s = &Service{
+		ctx:                ctx,
+		host:               h,
+		natManager:         natManager,
+		autonatDialer:      dialer,
+		pingDialer:         pingDialer,
+		handshakeService:   handshakeService,
+		libp2pPeerstore:    libp2pPeerstore,
+		metrics:            newMetrics(),
+		networkID:          networkID,
+		peers:              peerRegistry,
+		addressbook:        ab,
+		blocklist:          blocklist.NewBlocklist(storer),
+		logger:             logger,
+		tracer:             tracer,
+		connectionBreaker:  breaker.NewBreaker(breaker.Options{}), // use default options
+		ready:              make(chan struct{}),
+		halt:               make(chan struct{}),
+		lightNodes:         lightNodes,
+		HeadersRWTimeout:   o.HeadersRWTimeout,
+		autoNAT:            nil,
+		autoTLSCertManager: nil,
+		zapLogger:          nil,
+		// The wasmws transport dials both ws and wss (the browser performs TLS),
+		// so both are enabled for dial-filtering; only raw TCP is unavailable.
+		enabledTransports: map[bzz.TransportType]bool{
+			bzz.TransportTCP: false,
+			bzz.TransportWS:  true,
+			bzz.TransportWSS: true,
+		},
+		allowPrivateCIDRs: o.AllowPrivateCIDRs,
+		chequebookStorer:  o.ChequebookStorer,
+	}
+
+	peerRegistry.setDisconnecter(s)
+
+	s.lightNodeLimit = defaultLightNodeLimit
+	if o.LightNodeLimit > 0 {
+		s.lightNodeLimit = o.LightNodeLimit
+	}
+
+	// Construct protocols.
+	id := protocol.ID(p2p.NewSwarmStreamName(handshake.ProtocolName, handshake.ProtocolVersion, handshake.StreamName))
+	matcher, err := s.protocolSemverMatcher(id)
+	if err != nil {
+		return nil, fmt.Errorf("protocol version match %s: %w", id, err)
+	}
+
+	s.host.SetStreamHandlerMatch(id, matcher, s.handleIncoming)
+
+	connMetricNotify := newConnMetricNotify(s.metrics)
+	h.Network().Notify(peerRegistry) // update peer registry on network events
+	h.Network().Notify(connMetricNotify)
+
+	return s, nil
+}
